requirements.txt
# -----------------------------
# Core Utilities
# -----------------------------
requests>=2.31.0
kaggle
dvc>=3.0.0

# -----------------------------
# Numerical & Data Processing
# -----------------------------
numpy==1.26.4
pandas==2.1.4

# -----------------------------
# Data Validation
# -----------------------------
great-expectations==0.18.21

# -----------------------------
# Visualization (EDA)
# -----------------------------
matplotlib>=3.7.0,<3.9
seaborn>=0.12.2,<0.14

# -----------------------------
# Machine Learning & Deep Learning
# -----------------------------
scikit-learn==1.3.2
tensorflow>=2.15.0
pillow>=10.0.0
opencv-python>=4.8.0

# -----------------------------
# Experiment Tracking & Registry
# -----------------------------
mlflow>=2.14.1
sqlalchemy>=1.4.0

# -----------------------------
# Model Serving
# -----------------------------
flask>=2.3.0,<3.0

# -----------------------------
# Model Interoperability (ONNX)
# -----------------------------
onnx>=1.14.0
onnxruntime>=1.15.0
skl2onnx>=1.14.0
tf2onnx>=1.15.0

# -----------------------------
# Workflow Orchestration
# -----------------------------
prefect>=2.16.0

# -----------------------------
# Testing & Code Quality
# -----------------------------
pytest>=7.4.0
flake8>=6.1.0

# -----------------------------
# Monitoring & Metrics
# -----------------------------
prometheus-client>=0.19.0
prometheus-flask-exporter>=0.22.4


